% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.nflplotR_clear_cache}
\alias{.nflplotR_clear_cache}
\title{Clear nflplotR Cache}
\usage{
.nflplotR_clear_cache()
}
\value{
Invisibly \code{NULL}
}
\description{
This function clears the memoised cache of all functions memoised by \code{nflplotR}.
}
\examples{
.nflplotR_clear_cache()
}
