% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_lines.R
\name{geom_lines}
\alias{geom_lines}
\alias{geom_median_lines}
\alias{geom_mean_lines}
\title{ggplot2 Layer for Horizontal and Vertical Reference Lines}
\usage{
geom_median_lines(...)

geom_mean_lines(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[ggpath:geom_mean_lines]{ggpath::geom_mean_lines}}, \code{\link[ggpath:geom_median_lines]{ggpath::geom_median_lines}}
  \describe{
    \item{\code{mapping}}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}.}
    \item{\code{data}}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}
    \item{\code{na.rm}}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}
    \item{\code{show.legend}}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display. To include legend keys for all levels, even
when no data exists, use \code{TRUE}.  If \code{NA}, all levels are shown in legend,
but unobserved levels are omitted.}
    \item{\code{inherit.aes}}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behavior from
the default plot specification.}
  }}
}
\value{
A ggplot2 layer (\code{\link[ggplot2:layer]{ggplot2::layer()}}) that can be added to a plot
created with \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.
}
\description{
These geoms can be used to draw horizontal or vertical reference
lines in a ggplot. They use the data in the aesthetics \code{x0} and \code{y0}
to compute their \code{median} or \code{mean} and draw them as a line.
}
\details{
These functions have been outsourced to the ggpath package.
See \code{\link[ggpath:geom_median_lines]{ggpath::geom_median_lines}} and \code{\link[ggpath:geom_mean_lines]{ggpath::geom_mean_lines}} for details.
}
\section{Aesthetics}{


\code{geom_median_lines()} and \code{geom_mean_lines()} understand the following
aesthetics (at least one of the \code{x0} or \code{y0} aesthetics is required):
\describe{
\item{\code{x0}}{The variable for which to compute the median/mean that is drawn as vertical line.}
\item{\code{y0}}{The variable for which to compute the median/mean that is drawn as horizontal line.}
\item{\code{alpha = NA}}{The alpha channel, i.e. transparency level, as a numerical value between 0 and 1.}
\item{\code{color = "red"}}{The color of the drawn lines.}
\item{\code{linetype = 2}}{The linetype of the drawn lines.}
\item{\code{linewidth = 0.5}}{The width of the drawn lines.}
}

}

\examples{
library(ggplot2)

# inherit top level aesthetics
ggplot(mtcars, aes(x = disp, y = mpg, y0 = mpg, x0 = disp)) +
  geom_point() +
  geom_median_lines() +
  geom_mean_lines(color = "blue") +
  theme_minimal()

# draw horizontal line only
ggplot(mtcars, aes(x = disp, y = mpg, y0 = mpg)) +
  geom_point() +
  geom_median_lines() +
  geom_mean_lines(color = "blue") +
  theme_minimal()

# draw vertical line only
ggplot(mtcars, aes(x = disp, y = mpg, x0 = disp)) +
  geom_point() +
  geom_median_lines() +
  geom_mean_lines(color = "blue") +
  theme_minimal()

# choose your own value
ggplot(mtcars, aes(x = disp, y = mpg)) +
  geom_point() +
  geom_median_lines(x0 = 400, y0 = 15) +
  geom_mean_lines(x0 = 150, y0 = 30, color = "blue") +
  theme_minimal()
}
\seealso{
The underlying ggplot2 geoms \code{\link[ggplot2:geom_hline]{ggplot2::geom_hline}} and \code{\link[ggplot2:geom_vline]{ggplot2::geom_vline}}
}
