% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_nfl.R
\name{gt_render_image}
\alias{gt_render_image}
\title{Render 'gt' Table to Temporary png File}
\usage{
gt_render_image(gt_tbl, ...)
}
\arguments{
\item{gt_tbl}{An object of class \code{gt_tbl} usually created bt \code{\link[gt:gt]{gt::gt()}}}

\item{...}{Arguments passed on to \code{\link[webshot2:webshot]{webshot2::webshot()}} and \code{\link[=par]{par()}}.}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Saves a gt table to a temporary png image file and uses magick to render
tables in reproducible examples like \code{reprex::reprex()} or in package
function examples.
}
\examples{
\donttest{
# Don't run this on CRAN because an underlying dependency (chromote)
# keeps open connections which causes R CMD check errors.
# This code will run in CI workflows. To avoid problems in those workflows,
# you can set the environment variable "_R_CHECK_CONNECTIONS_LEFT_OPEN_" to
# "false"
if (Sys.getenv("NOT_CRAN") == "true") {
tbl <- gt::gt_preview(mtcars)
gt_render_image(tbl)
}
}
}
