% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfl_team_tiers.R
\name{nfl_team_tiers}
\alias{nfl_team_tiers}
\title{Create NFL Team Tiers}
\usage{
nfl_team_tiers(
  data,
  season = NULL,
  week = NULL,
  tier_desc = c(`1` = "Super Bowl", `2` = "Very Good", `3` = "Medium", `4` = "Bad", `5`
    = "What are they doing?"),
  presort = FALSE,
  alpha = 0.8,
  width = 0.075
)
}
\arguments{
\item{data}{A data frame that has to include the variables \code{tier_no} (the
number of the tier starting from the top tier no. 1) and \code{team_abbr} (the
team abbreviation). \code{team_abbr} should be one of \code{\link[=valid_team_names]{valid_team_names()}} and
the function tries to clean team names internally by calling
\code{\link[nflreadr:clean_team_abbrs]{nflreadr::clean_team_abbrs()}}.}

\item{season}{The season the team tiers are built for. This will be placed
in the title of the plot.}

\item{week}{The week inside \code{season} the team tiers are built for. This will
be placed in the title of the plot.}

\item{tier_desc}{A named vector consisting of the tier descriptions. The names
must equal the tier numbers from \code{tier_no}}

\item{presort}{If \code{FALSE} (the default) the function assumes that the teams
are already sorted within the tiers. Will otherwise sort alphabetically.}

\item{alpha}{The alpha channel of the logos, i.e. transparency level, as a
numerical value between 0 and 1.}

\item{width}{The desired width of the logo in \code{npc} (Normalised Parent Coordinates).}
}
\description{
This function sets up a ggplot to visualize NFL team tiers.
}
\examples{
\donttest{
library(ggplot2)
library(dplyr, warn.conflicts = FALSE)
team_abbr <- valid_team_names()
# remove conference logos from this example
team_abbr <- team_abbr[!team_abbr \%in\% c("AFC", "NFC")]

# Build the team tiers data frame
# This is completely random!
df <- data.frame(
  tier_no = sample(1:5, length(team_abbr), replace = TRUE),
  team_abbr = team_abbr
) \%>\%
  dplyr::group_by(tier_no) \%>\%
  dplyr::mutate(tier_rank = sample(1:n(), n()))

nfl_team_tiers(df, 2021, 4)
}
}
