% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_nfl_headshots.R, R/geom_nfl_logos.R,
%   R/geom_nfl_wordmarks.R, R/nflplotR-package.R
\docType{package}
\name{GeomNFLheads}
\alias{nflplotR}
\alias{GeomNFLheads}
\alias{GeomNFLlogo}
\alias{GeomNFLwordmark}
\alias{nflplotR-package}
\title{nflplotR: NFL Logo Plots in 'ggplot2' and 'gt'}
\usage{
GeomNFLheads

GeomNFLlogo

GeomNFLwordmark
}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A set of functions to visualize National Football League analysis in 'ggplot2' plots and 'gt' tables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nflplotr.nflverse.com}
  \item \url{https://github.com/nflverse/nflplotR}
  \item Report bugs at \url{https://github.com/nflverse/nflplotR/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Carl \email{mrcaseb@gmail.com}

Authors:
\itemize{
  \item Sebastian Carl \email{mrcaseb@gmail.com}
}

}
\keyword{internal}
