% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nflverse_sitrep}
\alias{nflverse_sitrep}
\title{Get a Situation Report on System, nflverse Package Versions and Dependencies}
\value{
Situation report of R and package/dependencies.
}
\description{
See \code{\link[nflreadr:nflverse_sitrep]{nflreadr::nflverse_sitrep}} for details.
}
