% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valid_team_names}
\alias{valid_team_names}
\title{Output Valid NFL Team Abbreviations}
\usage{
valid_team_names(exclude_duplicates = TRUE)
}
\arguments{
\item{exclude_duplicates}{If \code{TRUE} (the default) the list of valid team
abbreviations will exclude duplicates related to franchises that have been
moved}
}
\description{
Output Valid NFL Team Abbreviations
}
\examples{
# List valid team abbreviations excluding duplicates
valid_team_names()

# List valid team abbreviations excluding duplicates
valid_team_names(exclude_duplicates = FALSE)
}
