% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_nfl.R
\name{gt_render_image}
\alias{gt_render_image}
\title{Render gt Table to temporary png file}
\usage{
gt_render_image(gt_tbl, ...)
}
\arguments{
\item{gt_tbl}{An object of class \code{gt_tbl} usually created bt \code{\link[gt:gt]{gt::gt()}}}

\item{...}{Arguments passed on to \code{\link[webshot2:webshot]{webshot2::webshot()}}}
}
\value{
Returns \code{NULL} invisibly.
}
\description{
Saves a gt table to a temporary png image file and renders the image
with magick to be able to render tables in reproducible examples like
\code{\link[reprex:reprex]{reprex::reprex()}} or in package function examples.
}
\examples{
tbl <- gt::gt_preview(mtcars)
gt_render_image(tbl)
}
