% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nflplotR-package.R
\docType{package}
\name{nflplotR-package}
\alias{nflplotR}
\alias{nflplotR-package}
\title{nflplotR: Logo Plots in 'ggplot2'}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Sebastian Carl \email{mrcaseb@gmail.com}

}
\keyword{internal}
