% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nflplotR-package.R
\docType{package}
\name{nflplotR-package}
\alias{nflplotR}
\alias{nflplotR-package}
\title{nflplotR: NFL Logo Plots in 'ggplot2'}
\description{
A set of functions to visualize National Football
    League analysis in 'ggplot2'.
}
\author{
\strong{Maintainer}: Sebastian Carl \email{mrcaseb@gmail.com}

}
\keyword{internal}
