% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_from_path.R, R/geom_lines.R,
%   R/geom_nfl_headshots.R, R/geom_nfl_logos.R, R/geom_nfl_wordmarks.R,
%   R/nflplotR-package.R
\docType{data}
\name{GeomFromPath}
\alias{GeomFromPath}
\alias{GeomRefLines}
\alias{GeomNFLheads}
\alias{GeomNFLlogo}
\alias{GeomNFLwordmark}
\alias{nflplotR}
\alias{nflplotR-package}
\title{nflplotR: NFL Logo Plots in 'ggplot2'}
\format{
An object of class \code{GeomFromPath} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.

An object of class \code{GeomRefLines} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.

An object of class \code{GeomNFLheads} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.

An object of class \code{GeomNFLlogo} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.

An object of class \code{GeomNFLwordmark} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomFromPath

GeomRefLines

GeomNFLheads

GeomNFLlogo

GeomNFLwordmark
}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A set of functions to visualize National Football League analysis in 'ggplot2'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nflplotr.nflverse.com}
  \item \url{https://github.com/nflverse/nflplotR}
  \item Report bugs at \url{https://github.com/nflverse/nflplotR/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Carl \email{mrcaseb@gmail.com}

}
\keyword{datasets}
\keyword{internal}
