% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nflplotR-package.R
\docType{package}
\name{nflplotR-package}
\alias{nflplotR}
\alias{nflplotR-package}
\title{nflplotR: NFL Logo Plots in 'ggplot2'}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A set of functions to visualize National Football League analysis in 'ggplot2'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nflplotr.nflverse.com}
  \item \url{https://github.com/nflverse/nflplotR}
  \item Report bugs at \url{https://github.com/nflverse/nflplotR/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Carl \email{mrcaseb@gmail.com}

}
\keyword{internal}
